- add code review slash command for high-throughput Java system analysis - add build directory and code review reports to gitignore - add JUnit and Mockito test dependencies and configure mainClass in build.gradle - add dual-buffer architecture documentation and ClickHouse storage details to README - add REST API documentation with /health and /history endpoints - add CandleAggregatorTest unit test class - add HistoryRequest validation class for REST API parameter handling - refactor price and volume from float to double/long for precision - refactor wait strategy from BlockingWaitStrategy to YieldingWaitStrategy - add validation for timeframe configuration parsing - add graceful shutdown implementation in NonDriftingTimer - add /health endpoint to HistoryService - update Bash cat permission in Claude settings

+++ src/main/java/com/price/event/CandleEvent.java
@@ -16,9 +16,9 @@ public class CandleEvent extends TraceableEvent {
     private String instrument;
     private int timeframeMs;
     private long time;
-    private float open;
-    private float high;
-    private float low;
-    private float close;
-    private float volume;
+    private double open;
+    private double high;
+    private double low;
+    private double close;
+    private long volume;
 }
